feature: add payment details tabs
Since we're gonna have multiple payment methods, an ekegant way of doing that is by adding tabs
for the customer navigation.

This commit adds the skeleton of this tab structure inserted into the outer grid layout of the
page. Now we need to enhance the styling and finish the forms.